requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
yfinance==0.2.32
pandas==2.1.3
numpy==1.26.2
APScheduler==3.10.4
groq==0.4.1
binance-connector==3.7.0
feedparser==6.0.10
